requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==1.8.6
idna==2.7
PyGithub==1.40
PyJWT==1.6.4
pytz==2018.4
requests==2.19.1
urllib3==1.23
virtualenv==16.0.0
